+++ weblate/accounts/migrations/0020_remove_projects_dashboard.py
@@ -11,7 +11,7 @@ def remove_projects_dashboard(apps, schema_editor):
         # Profile.DASHBOARD_ALL
         dashboard_view=3
     ).update(
-        # Profile.DASHBOARD_SUBSCRIPTIONS
+        # Profile.DASHBOARD_WATCHED
         dashboard_view=1
     )