+++ docs/changes.rst
@@ -10,6 +10,7 @@ Released on ? 2016.
 * Extended import_json to be able to handle with existing components.
 * Added support for YAML files.
 * Project owners can now configure translation component and project details.
+* Use "Watched" instead of "Subscribed" projects.
 
 weblate 2.8
 -----------